backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.9.1
web3>=7.4.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
aiogram>=3.10.0
google-generativeai>=0.8.3
gunicorn>=21.2.0
redis[hiredis]>=5.0.0
sentry-sdk[fastapi]>=2.0.0
prometheus-fastapi-instrumentator>=7.0.0
